balance-sheet-rag/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.3.0
numpy>=1.24.3
python-dotenv==1.0.0
pandas>=2.0.3
scikit-learn>=1.3.0